Lift POI functions to App component
