+++ src/components/EditPage/EditPage.js
@@ -7,11 +7,8 @@ import POIListings from '../POIListings/POIListings';
 import { useParams } from 'react-router'
 
 
-function EditPage({addTrip, editTrip, trips, setTrips, checkTrip}) {
-  const [POIs, setPOIs] = useState('')
-  const [selectedPOI, setSelectedPOI] = useState('')
-  const [currTrip, setCurrTrip] = useState('')
-
+function EditPage({trips, POIs, setPOIs, selectedPOI, currTrip, setCurrTrip, highlightSelectedPOI, updateSelectedPOI, removePOI}) {
+  
   const { tripID } = useParams()
 
   useEffect(() => {
@@ -21,37 +18,6 @@ function EditPage({addTrip, editTrip, trips, setTrips, checkTrip}) {
       setPOIs(matchedTrip.destinations)
   }
   }, [tripID])
-
-  function highlightSelectedPOI(id) {
-    setSelectedPOI(POIs.find(poi => poi.id === id))
-  }
-
-  function updateSelectedPOI(id) {
-    const currPOI = POIs.find(poi => poi.id === id)
-    let updatedDestinations = currTrip.destinations
-    const isDupe = currTrip.destinations.some(destination => destination.id === id)
-    if (!isDupe) {
-      updatedDestinations = {
-        destinations: [...currTrip.destinations, currPOI]
-      }
-      const newTrip = {...currTrip, ...updatedDestinations}
-      setCurrTrip(newTrip) 
-      if (!trips[0] || checkTrip(newTrip.id)) {
-        addTrip(newTrip)
-      } else if (!checkTrip(newTrip.id)) {
-        editTrip(newTrip, currPOI)
-      } 
-    } 
-  }
-
-  function removePOI(id) {
-    const newTrips = [...trips]
-    const oldTrip = newTrips.find(trip => trip.id === currTrip.id)
-    const oldDestinationIndex = oldTrip.destinations.findIndex(destination => destination.id === id)
-    oldTrip.destinations.splice(oldDestinationIndex, 1)
-    setTrips(() => ([...newTrips]))
-    setCurrTrip(oldTrip)
-  }
 
   return ( 
 <>
