+++ src/components/ViewPage/ViewPage.js
@@ -5,11 +5,7 @@ import Map from '../Map/Map';
 import POIListings from '../POIListings/POIListings';
 import { useParams, useNavigate } from 'react-router'
 
-function ViewPage({addTrip, editTrip, trips, setTrips, checkTrip}) {
-  const [POIs, setPOIs] = useState('')
-  const [selectedPOI, setSelectedPOI] = useState('')
-  const [currTrip, setCurrTrip] = useState('')
-
+function ViewPage({trips, POIs, setPOIs, selectedPOI, currTrip, setCurrTrip, highlightSelectedPOI, updateSelectedPOI, removePOI}) {
   const { tripID } = useParams()
   let navigate = useNavigate()
 
@@ -21,36 +17,6 @@ function ViewPage({addTrip, editTrip, trips, setTrips, checkTrip}) {
   }
   }, [tripID])
 
-  function highlightSelectedPOI(id) {
-    setSelectedPOI(POIs.find(poi => poi.id === id))
-  }
-
-  function updateSelectedPOI(id) {
-    const currPOI = POIs.find(poi => poi.id === id)
-    let updatedDestinations = currTrip.destinations
-    const isDupe = currTrip.destinations.some(destination => destination.id === id)
-    if (!isDupe) {
-      updatedDestinations = {
-        destinations: [...currTrip.destinations, currPOI]
-      }
-      const newTrip = {...currTrip, ...updatedDestinations}
-      setCurrTrip(newTrip) 
-      if (!trips[0] || checkTrip(newTrip.id)) {
-        addTrip(newTrip)
-      } else if (!checkTrip(newTrip.id)) {
-        editTrip(newTrip, currPOI)
-      } 
-    } 
-  }
-
-  function removePOI(id) {
-    const newTrips = [...trips]
-    const oldTrip = newTrips.find(trip => trip.id === currTrip.id)
-    const oldDestinationIndex = oldTrip.destinations.findIndex(destination => destination.id === id)
-    oldTrip.destinations.splice(oldDestinationIndex, 1)
-    setTrips(() => ([...newTrips]))
-    setCurrTrip(oldTrip)
-  }
   return ( 
     <>
       <Nav />
